Refactored menu handling Added menu to select view type
