+++ asn1editor/wxPython/MainWindow.py
@@ -2,7 +2,6 @@ import os
 import sys
 import typing
 
-import asn1tools
 import wx
 import wx.svg
 
@@ -13,7 +12,7 @@ from asn1editor.PluginInterface import PluginInterface
 from asn1editor.wxPython import Environment
 from asn1editor.wxPython import WxPythonViewFactory
 from asn1editor.wxPython.FilePickerHandler import FilePickerHandler
-from asn1editor.wxPython.Resources import resource_path
+from asn1editor.wxPython.MenuHandler import MenuHandler
 from asn1editor.wxPython.SingleFileDropTarget import SingleFileDropTarget
 from asn1editor.wxPython.Styler import Styler
 from asn1editor.wxPython.treeView.TreeView import TreeView
@@ -25,18 +24,21 @@ class MainWindow(wx.Frame, PluginInterface):
 
         Environment.load()
 
-        self.__plugins = plugins
-        if self.__plugins is not None:
-            for plugin in self.__plugins:
+        if plugins is not None:
+            for plugin in plugins:
                 plugin.connect(self)
 
         self._status_bar = self.CreateStatusBar()
 
-        self.__create_menu()
+        self._menu_handler = MenuHandler(self, plugins)
+
+        self._menu_handler.build(self.load_spec, self.load_data_from_file, self.save_data_to_file)
+        self.Bind(wx.EVT_CLOSE, self.close, self)
 
         self.SetSize(wx.Size(Environment.settings.get('size', (500, 800))))
         self.Maximize(Environment.settings.get('maximized', True))
         self.SetPosition(wx.Point(Environment.settings.get('position', (0, 0))))
+        self._menu_handler.view_select.selected = Environment.settings.get('view', 1)
 
         self.__main_panel = wx.ScrolledWindow(self, style=wx.HSCROLL | wx.VSCROLL)
         self.__main_panel.SetScrollbars(15, 15, 50, 50)
@@ -53,130 +55,16 @@ class MainWindow(wx.Frame, PluginInterface):
 
         self.__progress_window: typing.Optional[wx.ProgressDialog] = None
 
-        self.bind_events()
-
         self.SetDropTarget(SingleFileDropTarget(self.__file_dropped))
 
         # noinspection SpellCheckingInspection
         sys.excepthook = self.__exception_handler
-
-    def __create_menu(self):
-        menu_bar = wx.MenuBar()
-        file_menu = wx.Menu()
-        self.__load_spec_item: wx.MenuItem = file_menu.Append(wx.ID_ANY, 'Open ASN.1 specification')
-        # noinspection PyArgumentList
-        image: wx.svg.SVGimage = wx.svg.SVGimage.CreateFromFile(resource_path('icons/open.svg'))
-        self.__load_spec_item.SetBitmap(image.ConvertToBitmap(width=16, height=16))
-        file_menu.AppendSeparator()
-        self.__load_data_item: wx.MenuItem = file_menu.Append(wx.ID_OPEN, 'Load encoded data')
-        # noinspection PyArgumentList
-        image: wx.svg.SVGimage = wx.svg.SVGimage.CreateFromFile(resource_path('icons/load_encoded.svg'))
-        self.__load_data_item.SetBitmap(image.ConvertToBitmap(width=16, height=16))
-        self.__load_data_item.Enable(False)
-        self.__save_data_item: wx.MenuItem = file_menu.Append(wx.ID_SAVE, 'Save encoded data')
-        # noinspection PyArgumentList
-        image: wx.svg.SVGimage = wx.svg.SVGimage.CreateFromFile(resource_path('icons/save_encoded.svg'))
-        self.__save_data_item.SetBitmap(image.ConvertToBitmap(width=16, height=16))
-        self.__save_data_item.Enable(False)
-        file_menu.AppendSeparator()
-        self.__exit_item = file_menu.Append(wx.ID_EXIT, 'Exit', 'Exit application')
-        # noinspection PyArgumentList
-        image: wx.svg.SVGimage = wx.svg.SVGimage.CreateFromFile(resource_path('icons/exit.svg'))
-        self.__exit_item.SetBitmap(image.ConvertToBitmap(width=16, height=16))
-        menu_bar.Append(file_menu, '&File')
-
-        view_menu = wx.Menu()
-        self.__auto_view: wx.MenuItem = view_menu.Append(1, 'Automatic', kind=wx.ITEM_CHECK)
-        self.__auto_view.Check(True)
-        self.__group_view: wx.MenuItem = view_menu.Append(2, 'Groups', kind=wx.ITEM_CHECK)
-        self.__tree_view: wx.MenuItem = view_menu.Append(3, 'Tree', kind=wx.ITEM_CHECK)
-        menu_bar.Append(view_menu, '&View')
-
-        help_menu = wx.Menu()
-        about_item = help_menu.Append(wx.ID_ABOUT, 'About')
-        self.Bind(wx.EVT_MENU, self.__about_item_event, about_item)
-        menu_bar.Append(help_menu, '&Help')
-
-        self.SetMenuBar(menu_bar)
-
-        toolbar: typing.Optional[wx.ToolBar] = None
-
-        if self.__plugins is not None:
-            for plugin_index, plugin in enumerate(self.__plugins):
-                plugin_menu = wx.Menu()
-                menus = plugin.get_menus()
-                for i, menu in enumerate(menus):
-                    if not len(menu[0]):
-                        plugin_menu.AppendSeparator()
-                    else:
-                        menu_item: wx.MenuItem = plugin_menu.Append(plugin_index * 1000 + i, menu[0])
-                        if menu[1] is not None:
-                            self.Bind(wx.EVT_MENU, self.__plugin_menu_event, menu_item)
-                        else:
-                            menu_item.Enable(False)
-
-                menu_bar.Append(plugin_menu, plugin.get_name())
-
-                tools = plugin.get_tools()
-                if len(tools):
-                    if toolbar is None:
-                        toolbar = self.CreateToolBar()
-                        toolbar.Bind(wx.EVT_TOOL, self.__tb_menu_event)
-                    else:
-                        toolbar.AddSeparator()
-
-                    for i, tool in enumerate(tools):
-                        if not len(tool[0]):
-                            toolbar.AddSeparator()
-                        else:
-                            bitmap = wx.Bitmap(tool[2])
-                            toolbar.AddTool(toolId=plugin_index * 1000 + i, label=tool[0], bitmap=bitmap, shortHelp=tool[1])
-
-        if toolbar is not None:
-            toolbar.Realize()
-
-    def __tb_menu_event(self, e):
-        menu_id = e.GetId()
-        plugin_index = menu_id // 1000
-        self.__plugins[plugin_index].get_tools()[menu_id % 1000][3]()
-
-    def __plugin_menu_event(self, e):
-        menu_id = e.GetId()
-        plugin_index = menu_id // 1000
-        self.__plugins[plugin_index].get_menus()[menu_id % 1000][1]()
 
     def __file_dropped(self, file_name: str):
         if self.__asn1_handler is not None:
             self.load_data_from_file(file_name)
         else:
             self.load_spec(file_name)
-
-    def bind_events(self):
-        def schema_dialog_constructor() -> wx.FileDialog:
-            return wx.FileDialog(self, "ASN.1 schema", wildcard="ASN.1 files (*.asn)|*.asn",
-                                 style=wx.FD_OPEN | wx.FD_FILE_MUST_EXIST | wx.FD_MULTIPLE)
-
-        def data_load_dialog_constructor() -> wx.FileDialog:
-            extensions = ';'.join(ASN1SpecHandler.get_extensions())
-            return wx.FileDialog(self, "ASN.1 encoded file", wildcard=f"ASN.1 encoded ({extensions})|{extensions}",
-                                 style=wx.FD_OPEN | wx.FD_FILE_MUST_EXIST)
-
-        def data_save_dialog_constructor() -> wx.FileDialog:
-            extensions = ';'.join(ASN1SpecHandler.get_extensions())
-            return wx.FileDialog(self, "ASN.1 encoded file", wildcard=f"ASN.1 encoded ({extensions})|{extensions}", style=wx.FD_SAVE)
-
-        self.Bind(wx.EVT_MENU, self.exit, self.__exit_item)
-
-        picker = FilePickerHandler(schema_dialog_constructor, self.load_spec)
-        self.Bind(wx.EVT_MENU, picker.on_menu_click, self.__load_spec_item)
-
-        picker = FilePickerHandler(data_load_dialog_constructor, self.load_data_from_file)
-        self.Bind(wx.EVT_MENU, picker.on_menu_click, self.__load_data_item)
-
-        picker = FilePickerHandler(data_save_dialog_constructor, self.save_data_to_file, True)
-        self.Bind(wx.EVT_MENU, picker.on_menu_click, self.__save_data_item)
-
-        self.Bind(wx.EVT_CLOSE, self.close, self)
 
     def load_spec(self, file_name: str, type_name: typing.Optional[str] = None):
         # Spec file loaded, compile it to show a selection of type names
@@ -207,11 +95,11 @@ class MainWindow(wx.Frame, PluginInterface):
             self.__view, self.__controller = self.__asn1_handler.create_mvc_for_type(self.__type_name, view_factory)
             sizer: wx.Sizer = self.__main_panel.GetSizer()
             sizer.Clear()
-            if False:
-                sizer.Add(self.__view.realize().get_sizer(), 0, wx.ALL | wx.EXPAND, 5)
-            else:
+            if self._menu_handler.view_select.selected == self._menu_handler.view_select.TREE:
                 tree_ctrl = TreeView.get_ctrl(self.__main_panel, self.__view.realize(), type_name)
                 sizer.Add(tree_ctrl, 0, wx.ALL | wx.EXPAND, 5)
+            else:
+                sizer.Add(self.__view.realize().get_sizer(), 0, wx.ALL | wx.EXPAND, 5)
 
             sizer.Layout()
 
@@ -221,8 +109,7 @@ class MainWindow(wx.Frame, PluginInterface):
 
             view_factory.thaw()
 
-            self.__load_data_item.Enable(True)
-            self.__save_data_item.Enable(True)
+            self._menu_handler.enable()
 
     def load_data_from_file(self, file_name: str):
         self.__controller.model_to_view(self.__asn1_handler.load_data_file(file_name))
@@ -326,30 +213,11 @@ class MainWindow(wx.Frame, PluginInterface):
         Environment.log_error(exception_str)
 
     # noinspection PyUnusedLocal
-    def __about_item_event(self, e):
-        del e
-        dialog = wx.MessageDialog(self, f'''asn1editor {asn1editor.__version__}
-
-Published under MIT License
-
-Copyright (c) 2020 Florian Fetz
-https://github.com/Futsch1/asn1editor
-
-Based on eerimoq's asn1tools, used in {asn1tools.version.__version__}
-https://github.com/eerimoq/asn1tools
-''', style=wx.ICON_INFORMATION | wx.OK, caption='About')
-        dialog.ShowModal()
-
-    # noinspection PyUnusedLocal
-    def exit(self, e: wx.Event):
-        del e
-        self.Close()
-
-    # noinspection PyUnusedLocal
     def close(self, e: wx.Event):
         Environment.settings['size'] = (self.GetSize().Get())
         Environment.settings['maximized'] = self.IsMaximized()
         Environment.settings['position'] = self.GetPosition().Get()
+        Environment.settings['view'] = self._menu_handler.view_select.selected
 
         Environment.save()
